cleanup and comments

+++ barretenberg/cpp/src/barretenberg/ultra_honk/oink_verifier.cpp
@@ -137,12 +137,11 @@ template <IsUltraOrMegaHonk Flavor> void OinkVerifier<Flavor>::execute_log_deriv
  */
 template <IsUltraOrMegaHonk Flavor> void OinkVerifier<Flavor>::execute_grand_product_computation_round()
 {
-    const FF public_input_delta =
-        compute_public_input_delta<Flavor>(public_inputs,
-                                           relation_parameters.beta,
-                                           relation_parameters.gamma,
-                                           verification_key->vk->log_circuit_size,
-                                           static_cast<size_t>(verification_key->vk->pub_inputs_offset));
+    const FF public_input_delta = compute_public_input_delta<Flavor>(public_inputs,
+                                                                     relation_parameters.beta,
+                                                                     relation_parameters.gamma,
+                                                                     verification_key->vk->log_circuit_size,
+                                                                     verification_key->vk->pub_inputs_offset);
 
     relation_parameters.public_input_delta = public_input_delta;
